backend/requirements.txt
fastapi
uvicorn
python-multipart
google-genai
pydantic
pdfplumber
Pillow
python-dotenv